Use dataclasses and mixins instead
