+++ darts/src/darts/legacy_pipeline/__init__.py
@@ -1,6 +1,6 @@
 """Legacy pipeline module."""
 
-from darts.legacy_pipeline.legacy import run_native_planet_pipeline as run_native_planet_pipeline
-from darts.legacy_pipeline.legacy import run_native_sentinel2_pipeline as run_native_sentinel2_pipeline
-from darts.legacy_pipeline.legacy_fast import run_native_planet_pipeline_fast as run_native_planet_pipeline_fast
-from darts.legacy_pipeline.legacy_fast import run_native_sentinel2_pipeline_fast as run_native_sentinel2_pipeline_fast
+from darts.legacy_pipeline.planet import run_native_planet_pipeline as run_native_planet_pipeline
+from darts.legacy_pipeline.planet_fast import run_native_planet_pipeline_fast as run_native_planet_pipeline_fast
+from darts.legacy_pipeline.s2 import run_native_sentinel2_pipeline as run_native_sentinel2_pipeline
+from darts.legacy_pipeline.s2_fast import run_native_sentinel2_pipeline_fast as run_native_sentinel2_pipeline_fast
